Added ruby file for user_account_app and moved necessary code to that file
