+++ bin/run.rb
@@ -1,4 +1,4 @@
 require_relative '../config/environment'
 
-run_main_menu
+# run_main_menu
